django/requirements.txt
appdirs==1.4.4
asgiref==3.7.2
astroid==3.1.0
beautifulsoup4==4.12.3
colorama==0.4.6
distlib==0.3.8
Django==4.2.11
django-admin-sortable==2.3
django-basicauth==0.5.3
django-bootstrap4==24.1
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-dbbackup==4.1.0
django-filter==24.1
django-fs-trumbowyg==0.1.4
django-js-asset==2.2.0
django-livereload-server==0.5.1
django-ordered-model==3.7.4
django-summernote==0.8.20.0
django-tinymce==3.7.1
Django-tinymce-filebrowser==0.2.1
djangorestframework==3.14.0
filelock==3.13.1
gunicorn==21.2.0
importlib-metadata==7.0.2
importlib-resources==6.3.0
isort==5.13.2
lazy-object-proxy==1.10.0
livereload==2.6.3
Markdown==3.6
mccabe==0.7.0
Pillow==10.2.0
psycopg2-binary==2.9.9
pylint==3.1.0
pytz==2024.1
six==1.16.0
sorl-thumbnail==12.10.0
soupsieve==2.5
sqlparse==0.4.4
toml==0.10.2
tornado==6.4
typed-ast==1.5.5
virtualenv==20.25.1
virtualenvwrapper-win==1.2.7
wrapt==1.16.0
zipp==3.18.1
